Move `RemoteControl.CodeMod.Ast.*` into `Ast`
`Lexical.RemoteControl.CodeMod.Ast` contained generally-useful functions for
parsing and creating zippers from documents. While it was located in
`RemoteControl`, it wasn't doing anything that required running on the remote
node.

Similarly, `Lexical.RemoteControl.CodeMod.Ast.Aliases` contains code for
resolving the available aliases at a given position and also has no
requirements to run on the remote node. It's been moved to
`Lexical.Ast.Aliases`.
